pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alpha-core"
version = "0.1.0"
description = "OFI+CVD 交易系统核心组件包"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "V13 Team", email = "team@v13.io"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "numpy>=1.20.0",
    "pytz>=2021.1",
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",
    "PyYAML>=6.0.0",
    "prometheus-client>=0.19.0",  # Prometheus指标支持（可选，未安装时使用降级实现）
    "python-binance>=1.0.19",  # Binance官方SDK，用于期货交易（解决签名问题）
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.5.0",
    "isort>=5.12.0",
]

[tool.setuptools]
# 使用自动发现，确保子包被打包
[tool.setuptools.packages.find]
where = ["src"]
include = ["alpha_core*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 120
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

